Updated test to use floats and doubles

+++ test/test_xfft.cpp
@@ -21,8 +21,8 @@ namespace xt
         size_t k = 2;
         size_t n = 8;
         size_t A = 10;
-        auto x = xt::linspace<double> (0, static_cast<double>(n - 1), n);
-        xt::xarray<double> y = A * xt::sin(2 * xt::numeric_constants<double>::PI * x * k / n);
+        auto x = xt::linspace<float> (0, static_cast<float>(n - 1), n);
+        xt::xarray<double> y = A * xt::sin(2 * xt::numeric_constants<float>::PI * x * k / n);
         auto res = xt::fft::ifft(y) / (n / 2);
         REQUIRE(A == doctest::Approx(std::abs(res(k))).epsilon(.0001));
     }
@@ -61,8 +61,8 @@ namespace xt
         size_t n = 15;
         size_t A = 1;
         size_t dim = 2;
-        auto x = xt::linspace<double>(0, n - 1, n) * xt::ones<double>({ dim, n });
-        xt::xarray<double> y = A * xt::sin(2 * xt::numeric_constants<double>::PI * x * k / n);
+        auto x = xt::linspace<float>(0, n - 1, n) * xt::ones<float>({ dim, n });
+        xt::xarray<double> y = A * xt::sin(2 * xt::numeric_constants<float>::PI * x * k / n);
         auto res = xt::fft::fft(y) / (n / 2.0);
         REQUIRE(A == doctest::Approx(std::abs(res(0, k))).epsilon(.0001));
         REQUIRE(A == doctest::Approx(std::abs(res(1, k))).epsilon(.0001));
